Move the test of test_client_on_distributed_vineyard.py to test_distributed.py
